decompose token_list.hpp
