+++ include/parser/check.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "tokens/tokens.hpp"
+#include "tokens/token_list.hpp"
 
 namespace check::symbol {
