+++ include/parser/parse_arith.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "tokens/token_list.hpp"
 #include "ast/expressions/expr.hpp"
 
 expr* parse_arith(token_list& Token_list);
