+++ include/parser/parse_logical.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "tokens/token_list.hpp"
 #include "ast/expressions/expr.hpp"
 
 expr* parse_logical_and(token_list& Token_list);
