+++ include/scanner/scanner.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
-#include "tokens/tokens.hpp"
+#include "tokens/tokens.hpp" //!
+#include "tokens/token_list.hpp"
 
 token_list scan(std::wstring filepath);
